requirements.txt
gunicorn
flask
python-binance